chore: return chain interact node error code

+++ wallet-api/src/error/business/chain_node.rs
@@ -5,7 +5,7 @@ pub enum ChainNodeError {
 }
 
 impl ChainNodeError {
-    pub(crate) fn get_status_code(&self) -> u32 {
+    pub(crate) fn get_status_code(&self) -> i64 {
         match self {
             ChainNodeError::ChainNotFound => 4201,
         }
