+++ wallet-api/src/error/business/coin.rs
@@ -7,7 +7,7 @@ pub enum CoinError {
 }
 
 impl CoinError {
-    pub(crate) fn get_status_code(&self) -> u32 {
+    pub(crate) fn get_status_code(&self) -> i64 {
         match self {
             CoinError::NotFound(_) => 4001,
             CoinError::InvalidContractAddress(_) => 4002,
