+++ wallet-api/src/error/business/exchange_rate.rs
@@ -5,7 +5,7 @@ pub enum ExchangeRate {
 }
 
 impl ExchangeRate {
-    pub(crate) fn get_status_code(&self) -> u32 {
+    pub(crate) fn get_status_code(&self) -> i64 {
         match self {
             ExchangeRate::NotFound => 4101,
         }
